// src/types.d.ts
export interface Options {
  label: string;
  value: string | number;
  subOptions?: Options[];
  description?: string;
  data?: any;
}

export interface MaterialEquipmentObj {
  code: string;
  uuid: string;
  id: number;
  quantity: number | string;
  size?: string;
  unit: string;
  name: string;
  price?: number | string;
  equipmentId?: string;
  equipmentAmount?: string | number;
  unitPrice?: string | number;
  availableStock?: number;
  isActive: boolean;
  isUpdate: boolean;
  isEdited: boolean;

  materialSetUuid?: string;
  originalId?: number;
}
// ...
export interface Execution {
  id: string;
  workOrderId: string;
  startDateTime: string | Date | undefined;
  endDateTime: string | Date | undefined;
  note: string;
  username: string;
  userType: string;
  executionDetail: string;
  createdAt: string;
  serviceSpecificData: ServiceSpecificData;
  images: UploaddedImage[];
  attachments: UploadedFile[];
  startWorkDate?: string;
  endWorkDate?: string;
  latitude?: number;
  longitude?: number;
  executionNote?: string;
  serviceSatisfaction?: number;
  satisfactionComment?: string;
  readonly recorderName?: string;
  readonly recorderPosition?: string;
  readonly recorderPhoneNumber?: string;
  recorderSignature?: string;
  customerSignature?: string;
}
// ...
export type WorkOrderObj = {
  id: string;
  requestCode: string;
  workOrderNo: string;
  customerName: string;
  customerRequestNo?: string;
  customerMobileNo: string;
  customerAddress: string;
  customerEmail: string;
  customerBp: string;
  customerCa: string;
  customerLatitude: number;
  customerLongitude: number;
  requestServiceDetail: RequestServiceDetail | string;
  serviceSpecificData?: ServiceSpecificData;
  businessTypeId: string;
  businessTypeName: string;
  assignees: Assignee[];
  participants: Assignee[];
  appointmentDate: Date | string | undefined;
  sapProcessCreatedDate: Date | string | undefined;
  workOrderStatusCode: string;
  workOrderStatusName: string;
  statusCode: string;
  isSurvey: boolean;
  equipments: MaterialEquipmentObj[];
  priority: number;
  workDescription: string;
  mainWorkCenterId: string;
  mainWorkCenterName: string;
  costCenter: string;
  peaOffice: string;
  peaOfficeCode?: string;
  peaOfficeName?: string;
  officePlant: string;
  peaNameFull: string;
  mainWorkOrder: MainWorkOrder;
  results: {
    id: string;
    workOrderNo: string;
  }[];
  executionNote: string;
  latitude: number;
  longitude: number;
  startWorkDate: Date | undefined | string;
  endWorkDate: Date | undefined | string;
  execution: Execution;
  images: number[];
  attachments: number[];
  serviceSatisfaction: number;
  satisfactionComment: string;
  customerSignatureBase64: string;
  recorderName: string;
  recorderPosition: string;
  recorderPhoneNumber: string;
  recorderSignatureBase64: string;
  workTypeName: string;
  workType: string;
  requestSapNo: string | number;
  sapOrderNo: string;
  serviceGroupName: string;
  serviceName: string;
  createdAt: string;
  requestChannel: string;
  requestDate: string;
  workOrderType: string;
  defaultActivityPmCode: string;

  sheetType: sting;
  serviceType: string;
  serviceTypes: string[];
  organization: string[];
  order_no: string | number;
  request_no: string | number;
  request_sap_no: string | number;
  request_type: string;
  request_status: string;
  payment_received_date: string;
  pea_office: string;
  plant_code: string;
  cost_center: string;
  customer_info: Customer;
  electrical: Electrical[];
  // equipments: Equipments[];
  workers: WorkerObj[];
  insulators: Insulator[];
  // transformer: Transformer[];
  meterequipment: MeterEquipment[];
  materialEquipments?: MaterialEquipmentObj[];
  workCategory: string
  activityPmId: string
  isSms: boolean
  sapStatusCodes?: string
};
// ...
export interface MainWorkOrder {
  id: string;
  peaOffice: string;
  workOrderNo: string;
  officePlant: string;
  costCenter: string;
  peaNameFull: string;
  serviceName: string;
}

interface RequestServiceItem {
  item_id: string;
  quantity: number | string;
  isUpdate?: boolean;
  isEdited?: boolean;
}

type Assignee = {
  id: number;
  Id?: number;
  index?: number;
  name: string;
  username: string;
  userType: string;
  workCenterId: string;
  workUnit: string;
  workActivityTypeId: string;
  workHours: number | string;
  workOrderId?: string;
  sequenceNo?: number;
  isRead?: boolean;
  isEquipmentResponsible?: boolean;
  isUpdate?: boolean;
  isEdited?: boolean;
  activity: string;
  mainWorkCenter: string;
  startDatetime: string | Date | undefined;
  endDatetime: string | Date | undefined;
  workcenterCode?: string;
};

type Customer = {
  customer_name: string;
  customer_mobile_no: string;
  customer_address: string;
  customer_bp: string;
  customer_ca: string;
  customer_latitude: string;
  customer_longitude: string;
  customer_email: string;
};

type Electrical = {
  id: number;
  name: string;
  size?: string;
  quantity: number;
  isUpdate?: boolean;
  isEdited: boolean;
};

type Equipments = {
  id: number;
  name: string;
  equipmentId: string;
  equipmentAmount: string | number;
  unitPrice: number;
  totalPrice: number;
  isUpdate?: boolean;
  isEdited: boolean;
};

export type ElectricalEquipment = {
  id: number;
  name: string;
  quantity: number;
  isUpdate?: boolean;
  isEdited: boolean;
};

export type Transformer = {
  id: number;
  title: string;
  phase: string;
  type: string;
  serial: string;
  size: string;
  voltage: string;
  isUpdate?: boolean;
  isEdited?: boolean;
};

export type Insulator = {
  id: number;
  item_title: string;
  quantity: number;
  isUpdate?: boolean;
  isEdited?: boolean;
};

interface MeterEquipmentServiceItem {
  item_id: string;
  item_size: number;
  quantity: number;
  price: number;
}

export type MeterEquipment = {
  id: number | string;
  equipment_id?: string;
  equipment_name: string;
  size: string | number;
  quantity: number;
  price?: number;
  isUpdate?: boolean;
  isEdited?: boolean;
};
// ...
type WorkerObj = {
  isUpdate?: boolean;
  isEdited: boolean;
  group: string;
  operation_center: string;
  id: number;
  hours: number;
  unit: string;
  worker_id?: string;
  worker: string;
  event: string;
  group_worker: string;
};
// ...
export type ServiceRequest = {
  id: string;
  workOrderNo: string;
  requestNo: string;
  serviceId: string;
  requestCode: string;
  serviceType: string;
  createdDate: string;
  createdAt: string;
  customerName: string;
  statusCode: string;
  customerRequestNo: string;
  customerAddress: string;
  customerMobileNo: string;
  serviceGroupName: string;
  serviceName: string;
  status: string;
  statusDescription: string;
  workOrderParentId: string;
  costCenter: string;
  peaOffice: string;
  officePlant: string;
  peaNameFull: string;
  sapOrderNo: string;
  defaultActivityPmCode: string;
};

export type BusinessTypeObj = {
  id: string;
  name: string;
};

export type Event = {
  id: number;
  ktext: string;
  lstar: string;
};

export type MainWorkCenter = {
  Id: number;
  Ktext: string;
  Arbpl: string;
};

export interface ResponsiblePersonObj {
  id?: number | string;
  username: string;
  isUpdate: boolean;
  isEdited: boolean;
}

export interface RequestServiceDetail
  extends InverterData,
    SiteSurveyAirCondition {
  businessTypeId?: string;
  package_id?: string;
  transformer_voltage?: string;
  request_type?: string;
  service_type?: string;
  items: RequestServiceItem[] | Transformer[] | Insulator[];
  renewable_source_id: string;
  renewable_type_id: string;
  source_other: string;
  year: Date | undefined;
  num_years: string | number;
  start_date?: Date | undefined;
  end_date?: Date | undefined;
  days?: number;
  transformer?: TransFormerS315[];
}
// ...
export interface TransFormerS315 {
  id?: string | number;
  transformer_size: string;
  transformer_qty: number;
  isUpdate?: boolean;
  isEdited?: boolean;
}

export interface InverterData {
  distanceSolarToInverter: number;
  distanceInverterToPanel: number;
  connectionType: "LAN" | "WiFi" | "other";
  otherConnectionDetail: string;
}

export interface SiteSurveyAirCondition {
  refrigerantPipeDistance: number;
  dcCableDistance: number;
  acCableDistance: number;
  houseFrontDirection: string;
}

export interface WorkOrderCreateItem {
  serviceId: string;
  workOrderType: "single" | "bulk";
  requestCode?: string;
  customerRequestNo?: string | null;
  workOrderParentId?: string;
  sapOrderNo?: string;
  serviceTypes?: string[];
  serviceSpecificData?: ServiceSpecificData;
  selectedPeaOffices?: string[];
}

export interface BulkWorkOrderCreateItem {
  bulkType: string;
  customerRequestNo?: string;
  existingParentWorkOrderId?: string;
  mainWorkOrder?: {
    serviceId: string;
  };
  selectedPeaOffices: string[];
}
// ...
export interface BusinessTypeOnlyData {
  businessTypeId: string;
}

export type ServiceSpecificData =
  | BusinessTypeOnlyData
  | S301ServiceData
  | S305ServiceData
  | S307ServiceData
  | S308ServiceData
  | S312ServiceData
  | S314ServiceData
  | S315ServiceData
  | S316ServiceData
  | S318ServiceData
  | S322ServiceData
  | S329ServiceData
  | S332SolarAirData
  | S332SolarBatteryData;

// S301: ขอซ่อมแซมอุปกรณ์ไฟฟ้า
export interface S301ServiceData {
  equipments: Array<S301EquipmentServiceData>;
}

export interface S301EquipmentServiceData {
  equipmentTypeId: string;
  amount: number;
  isUpdate: boolean;
  isEdited: boolean;
}

// S305: ขอบำรุงรักษาหม้อแปลงไฟฟ้า
export interface S305ServiceData {
  requestServiceTypeId: string;
  requestServiceId: string;
  transformers: Array<S305TransformerServiceData>;
}

export interface S305TransformerServiceData {
  transformerBrandId: string;
  transformerPhaseId: string;
  transformerTypeId: string;
  transformerSerial?: string;
  transformerSize?: string;
  transformerVoltage?: string;
  isUpdate: boolean;
  isEdited: boolean;
}

// S307: ขอแก้ไขและบำรุงรักษาระบบจำหน่ายโดย Hotline
export interface S307ServiceData {
  businessTypeId: string;
  voltageId: string;
}

// S308: ขอตรวจสอบระบบไฟฟ้าพร้อมออกใบรับรอง
export interface S308ServiceData {
  businessTypeId: string;
  transformerCapacityKw: number;
  transformers: Array<{
    transformerBrandId: string;
    transformerPhaseId: string;
    transformerTypeId: string;
    transformerSerial: string;
    transformerSize: string;
    transformerVoltage: string;
  }>;
}

// S312: ขอทดสอบอุปกรณ์ไฟฟ้า
export interface S312ServiceData {
  businessTypeId: string;
  voltageId: string;
  equipments: Array<S301EquipmentServiceData>;
}

// S314: ขอเช่าฉนวนครอบสายไฟฟ้า
export interface S314ServiceData {
  rentalStartDate: string | Date | undefined;
  rentalEndDate: string | Date | undefined;
  rentalDays: number;
  cableInsulators: Array<S314CableServiceData>;
}

export interface S314CableServiceData {
  cableInsulator: string;
  amount: number;
  isUpdate: boolean;
  isEdited: boolean;
}

// S315: ขอเช่าหม้อแปลง
export interface S315ServiceData {
  businessTypeId: string;
  rentalStartDate: string;
  rentalEndDate: string;
  rentalDays: number;
  transformers: Array<S315TransformerServiceData>;
}

export interface S315TransformerServiceData {
  transformerCapacity: string;
  amount: number;
  isUpdate: boolean;
  isEdited: boolean;
}

// S316: ขอเช่าเครื่องกำเนิดไฟฟ้า
export interface S316ServiceData {
  rentalStartDate: string;
  rentalEndDate: string;
  rentalDays: number;
  generatorStartDate: string | Date | undefined;
  generatorEndDate: string | Date | undefined;
  generatorStartTime: string;
  generatorEndTime: string;
  generators: Array<S316GeneratorServiceData>;
}

export interface S316GeneratorServiceData {
  transformerCapacityId: string;
  peaOfficeId: string;
  amount: number;
}

export interface S316GeneratorElectric {
  kwSize: string;
  peaName: string;
  telNumber: string;
  distanceKm: number;
}

// S318: ขอซื้อมิเตอร์/อุปกรณ์ไฟฟ้า
export interface S318ServiceData {
  equipments: Array<S318EquipmentServiceData>;
}

export interface S318EquipmentServiceData {
  equipmentId: string;
  capacity: string;
  amount: number;
  isUpdate: boolean;
  isEdited: boolean;
  price: string | number;
}

// S322: ขอตรวจสอบและบำรุงรักษาระบบไฟฟ้า แบบครบวงจร (Package)
export interface S322ServiceData {
  businessTypeId: string;
  packageId: string;
  packageTitle: string;
}

// S329: ขอซื้อขายใบรับรองการผลิตพลังงานหมุนเวียน
export interface S329ServiceData {
  businessTypeId: string;
  renewableType: string;
  renewableSource: string;
  year: string | Date | undefined;
  qty: number;
}

// S332 Solar Air
export interface S332SolarAirData {
  pipeLengthRefrigerantM?: number;
  cableLengthDcM?: number;
  cableLengthAcM?: number;
  houseFacingDirection?: string;
  frontHouseMediaId?: number;
  roofInstallationMediaId?: number;
  outdoorUnitLocationMediaId?: number;
  indoorUnitLocationMediaId?: number;
  acWiringPathMediaId?: number;
  dcWiringPathMediaId?: number;

  // ฟิลด์ที่ backend ส่งกลับ (เมื่อ GET เท่านั้น)
  //execution
  frontHouse?: UploaddedImage;
  acModel?: UploaddedImage;
  solarPanelModel?: UploaddedImage;
  mountingStructure?: UploaddedImage;
  solarPanelInstallation?: UploaddedImage;
  indoorOutdoorUnits?: UploaddedImage;
  pipingAndDrain?: UploaddedImage;
  acBreakerWiring?: UploaddedImage;
  dcBreakerWiring?: UploaddedImage;
  warrantyDoc?: UploaddedImage;
  groundingSystem?: UploaddedImage;

  //survey
  roofInstallationUrl?: string;
  outdoorUnitLocationUrl?: string;
  indoorUnitLocationUrl?: string;
  acWiringPathUrl?: string;
  dcWiringPathUrl?: string;
}
// ...
interface UploaddedImage {
  id: number | string;
  file_id: string;
  name: string;
  fileName: string;
  originalName: string;
  size: number;
  fileSize: number;
  url: string;
  file: File;
  uploadDate?: Date;
  isUploading?: boolean;
  isOffline?: boolean;
}

interface UploadedFile {
  id: number | string;
  name: string;
  fileName?: string;
  originalName?: string;
  size: number;
  fileSize?: number;
  url: string;
  file: File;
  uploadDate?: Date;
  isUploading?: boolean;
  isOffline?: boolean;
  updatedAt?: string
}